[build] Revamp build process to build deps with main project.
The new build process builds all dependencies together with the
main project. Each dependency is processed in its own cmake file
(see cmake/extern directory). Depending on the build options,
the dependencies are build in the new ./extern directory. Patches
to external dependencies are also stored there. Dependencies are
not committed to git but downloaded and processed with cmake's
ExternalProject_Add() facilities.

This also introduces the option to build a static Scan Tailor
executable that does not need to be shipped with all the dynamic
libraries. This is the default mode when building all the
dependencies.